chore(mme): Split spgw_procedures_test
