+++ lte/gateway/c/core/oai/test/spgw_task/CMakeLists.txt
@@ -28,6 +28,8 @@ add_library(SPGW_TASK_TEST_LIB
     mock_spgw_op.h
     mock_spgw_op.cpp
     spgw_test.cpp
+    spgw_procedures_test_fixture.h
+    spgw_procedures_test_fixture.cpp
     )
 
 link_directories(/usr/src/googletest/googlemock/lib/)
@@ -42,7 +44,9 @@ target_link_libraries(SPGW_TASK_TEST_LIB
 
 set_target_properties(SPGW_TASK_TEST_LIB PROPERTIES LINKER_LANGUAGE CXX)
 
-foreach (sgw_test spgw_service_impl spgw_state_converter spgw_procedures pgw_pco spgw_procedures_with_injected_state)
+foreach (sgw_test spgw_service_impl spgw_state_converter 
+         spgw_procedures spgw_procedures_dedicated_bearer spgw_procedures_session
+         pgw_pco spgw_procedures_with_injected_state)
   add_executable(${sgw_test}_test test_${sgw_test}.cpp)
   target_link_libraries(${sgw_test}_test SPGW_TASK_TEST_LIB TASK_GRPC_SERVICE)
   add_test(test_${sgw_test} ${sgw_test}_test)
